backend/requirements.txt
# File: backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
PyMuPDF